[test/#130]: 회원정보 수정 테스트 코드를 기존 테스트 파일로 이동 및 중복 파일 삭제
